Move plugin creation behind main IPC
